Added option to show action globally, regardless of blocked portlets.
Added controlpanel and settings for this.
The timeout is now always the time since the first visit of a page with this portlet.
We save the start time in the cookie.

If you leave the site before viewing the action,
and you come back after more than an hour,
the timeout is reset: we start counting fresh.